+++ CHANGELOG.md
@@ -4,6 +4,7 @@ _See also: [the official libGDX changelog](https://github.com/libgdx/libgdx/blob
 
 - **[UPDATE]** Updated to Kotlin 1.7.0.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.6.2.
+- **[FEATURE]** (`ktx-actors`) Added `Actor.onExit` and `Actor.onEnter` extension methods that attach `InputListener` to an `Actor`
 
 #### 1.11.0-rc1